refactor(grid): [grid] inline design config
