+++ packages/design/aurora/src/grid/index.ts
@@ -1,6 +1,13 @@
 import { iconChevronUp, iconChevronDown } from '@opentiny/vue-icon'
 
 export default {
+  validConfig: {
+    icon: ''
+  },
+  minWidth: 40,
+  treeConfig: {
+    renderIcon: false
+  },
   icons: {
     // aurora不需要默认排序图标
     sortDefault: null,
