--- packages/design/smb/src/grid/index.ts
@@ -1,15 +1,0 @@
-import { iconError, iconExpand, iconPutAway } from '@opentiny/vue-icon'
-
-export default {
-  validConfig: {
-    icon: iconError()
-  },
-  minWidth: 72,
-  treeConfig: {
-    renderIcon(h, { active }) {
-      const IconExpand = iconExpand()
-      const IconPutAway = iconPutAway()
-      return active ? h(IconExpand) : h(IconPutAway)
-    }
-  }
-}
